rename OpticalFlowSensor
